[core] split `util` module in `utils` package; keeping backward compatibility
